Went through the code in detail. Added various constants and changed some things around. Simplified looping code. Removed or commented out obsolete code in some places. Changed the way the pulseOff timer is attached to prevent it from triggered unecessarily.

+++ EFICode/Sensors.cpp
@@ -5,8 +5,8 @@
 
 const double voltageConversion = .0049;
 
-int Controller::getRPM (int timePassed, int rev) {
-  return (60000000 * rev) / ((double) timePassed);
+int Controller::getRPM (long int timePassed, int rev) {
+  return (60000000.0 * rev) / (timePassed);
 }
 
 //TPS Measurement
